# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AssistantConversation {
  id: number;
  userId: string;
  title: string;
  createdAt: string;
}

export interface AssistantConversationInput {
  title?: string;
}

export type AssistantMessageRole = typeof AssistantMessageRole[keyof typeof AssistantMessageRole];


export const AssistantMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AssistantMessage {
  id: number;
  conversationId: number;
  role: AssistantMessageRole;
  content: string;
  toolsUsed?: string[] | null;
  createdAt: string;
}

export interface AssistantConversationWithMessages {
  id: number;
  userId: string;
  title: string;
  createdAt: string;
  messages: AssistantMessage[];
}

/**
 * Structured vehicle specs used for deterministic trail-fit checks. Omitted fields are treated as unknown/stock.
 */
export interface AssistantVehicleProfile {
  liftIn?: number;
  tireDiameterIn?: number;
  hasLockers?: boolean;
  hasLowRange?: boolean;
}

export interface AssistantMessageInput {
  content: string;
  vehicleProfile?: AssistantVehicleProfile;
}

export interface AssistantError {
  error: string;
}

/**
 * Firebase UID of the requesting user (trusted at face value in Phase 1)
 */
export type UserIdHeaderParameter = string;