# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: assistant
    description: AI trip assistant chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /assistant/conversations:
    get:
      operationId: listAssistantConversations
      tags: [assistant]
      summary: List all conversations for the current user
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantConversation"
    post:
      operationId: createAssistantConversation
      tags: [assistant]
      summary: Create a new conversation
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversation"
  /assistant/conversations/{id}:
    get:
      operationId: getAssistantConversation
      tags: [assistant]
      summary: Get a conversation with its messages
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
    delete:
      operationId: deleteAssistantConversation
      tags: [assistant]
      summary: Delete a conversation
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
  /assistant/conversations/{id}/messages:
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Send a message and receive an AI response (SSE stream)
      description: >
        Runs a multi-turn tool-calling agent loop against Claude and streams
        the result back as Server-Sent Events. Not consumable via the
        generated React Query hook — clients must use `expo/fetch` (native)
        or the browser `fetch` streaming body directly.
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantError"
components:
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: Firebase UID of the requesting user (trusted at face value in Phase 1)
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AssistantConversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - createdAt
    AssistantConversationInput:
      type: object
      properties:
        title:
          type: string
    AssistantMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        toolsUsed:
          type: array
          items:
            type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    AssistantConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AssistantMessage"
      required:
        - id
        - userId
        - title
        - createdAt
        - messages
    AssistantVehicleProfile:
      type: object
      description: Structured vehicle specs used for deterministic trail-fit checks. Omitted fields are treated as unknown/stock.
      properties:
        liftIn:
          type: number
        tireDiameterIn:
          type: number
        hasLockers:
          type: boolean
        hasLowRange:
          type: boolean
    AssistantMessageInput:
      type: object
      properties:
        content:
          type: string
        vehicleProfile:
          $ref: "#/components/schemas/AssistantVehicleProfile"
      required:
        - content
    AssistantError:
      type: object
      properties:
        error:
          type: string
      required:
        - error