# lib/api-zod/src/generated/types/assistantMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantVehicleProfile } from './assistantVehicleProfile';

export interface AssistantMessageInput {
  content: string;
  vehicleProfile?: AssistantVehicleProfile;
}