# lib/api-zod/src/generated/types/assistantVehicleProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Structured vehicle specs used for deterministic trail-fit checks. Omitted fields are treated as unknown/stock.
 */
export interface AssistantVehicleProfile {
  liftIn?: number;
  tireDiameterIn?: number;
  hasLockers?: boolean;
  hasLowRange?: boolean;
}